+++ energyquantified/time/timezone.py
@@ -1,6 +1,8 @@
 import pytz
 import tzlocal
 
+from ._timezone_gas import _build_europe_gas_day_tzinfo
+
 
 # Look up the most used time zones in the European power markets
 # from the pytz library
@@ -10,6 +12,7 @@ CET = pytz.timezone("CET")
 WET = pytz.timezone("WET")
 EET = pytz.timezone("EET")
 TRT = pytz.timezone("Europe/Istanbul")
+GAS_DAY = _build_europe_gas_day_tzinfo()
 
 
 def local_tz():